+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv (development version)
 
+* Fixed an issue where `renv` could fail to parse remote references
+  containing an `@` character. (#2135)
+
 * Fixed an issue where `renv::sysreqs(distro = <...>)` could fail when
   passed an alternate distribution without a version specifier. (#2105)